chore: cleanup middlewares, move defaults.
