+++ server.go
@@ -25,13 +25,6 @@ type Server struct {
 	Logger      *zap.Logger
 }
 
-func configureRouter() {
-	router := httprouter.New()
-	router.OPTIONS("*", func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-
-	})
-}
-
 // NewServer creates a new server instance
 func NewServer(config *ServerConfig) *Server {
 	if config.Logger == nil {
@@ -44,9 +37,6 @@ func NewServer(config *ServerConfig) *Server {
 		Config:      config,
 		Logger:      config.Logger,
 	}
-
-	newSTKServer.Use(SecurityHeaders)
-	newSTKServer.Use(CORS)
 
 	return newSTKServer
 }
